Use a real Kafka for all Kafka tests
Mocking out all the pieces of Kafka is weirdly annoying and causes
a few odd issues, and it also makes it harder to write tests without
arbitrary delays. Switch the remaining end-to-end tests to use a
real Kafka container and merge the test fixtures, deleting all the
mock ones. Test a few more things with real expected status messages.

Remove the various injection points in the code that the test suite
used to try to wire mocks into all the necessary places.
