+++ tests/data/status/simple-error.json
@@ -1,0 +1,14 @@
+{
+  "job_id": "uws123",
+  "execution_id": "1",
+  "timestamp": 1743540791,
+  "status": "ERROR",
+  "error": {
+    "code": "backend_error",
+    "message": "Qserv request failed: Some error"
+  },
+  "metadata": {
+    "query": "SELECT TOP 10 * FROM table",
+    "database": "dp1"
+  }
+}
