+++ tests/kafka/query_test.py
@@ -11,6 +11,7 @@ from aiokafka import AIOKafkaConsumer
 from asgi_lifespan import LifespanManager
 from fastapi import FastAPI
 from faststream.kafka import KafkaBroker
+from httpx import Response
 from safir.datetime import current_datetime
 from testcontainers.redis import RedisContainer
 
@@ -24,6 +25,7 @@ from qservkafka.models.state import Query
 
 from ..support.arq import run_arq_jobs
 from ..support.data import (
+    read_test_job_cancel,
     read_test_job_run,
     read_test_job_run_json,
     read_test_job_status,
@@ -261,6 +263,58 @@ async def test_failure(
 
 @pytest.mark.asyncio
 @pytest.mark.timeout(60)
+async def test_qserv_error(
+    *,
+    app: FastAPI,
+    kafka_broker: KafkaBroker,
+    kafka_status_consumer: AIOKafkaConsumer,
+    mock_qserv: MockQserv,
+    redis: RedisContainer,
+) -> None:
+    """Test proper handling of an API error getting completed job status.
+
+    An earlier version of the Qserv Kafka bridge erroneously didn't stop
+    processing when the API request failed.
+    """
+    async with LifespanManager(app):
+        factory = context_dependency.create_factory()
+
+        await start_query(kafka_broker, "simple")
+        status = await wait_for_status(kafka_status_consumer, "simple-started")
+        assert status.query_info
+        start_time = status.query_info.start_time
+
+        mock_qserv.set_status_response(
+            Response(
+                200,
+                json={"success": 0, "error": "Some error"},
+            )
+        )
+        await mock_qserv.update_status(
+            1,
+            AsyncQueryStatus(
+                query_id=1,
+                status=AsyncQueryPhase.COMPLETED,
+                total_chunks=10,
+                completed_chunks=10,
+                collected_bytes=250,
+                query_begin=start_time,
+                last_update=datetime.now(tz=UTC),
+            ),
+        )
+        await wait_for_dispatch(factory, 1)
+
+        # Run the background tsk queue.
+        assert await run_arq_jobs() == 1
+        status = await wait_for_status(kafka_status_consumer, "simple-error")
+
+    # Ensure all query state has been deleted.
+    redis_client = redis.get_client()
+    assert set(redis_client.scan_iter("query:*")) == set()
+
+
+@pytest.mark.asyncio
+@pytest.mark.timeout(60)
 async def test_missing_executing(
     *,
     app: FastAPI,
@@ -293,6 +347,85 @@ async def test_missing_executing(
     assert raw_query
     query = Query.model_validate(json.loads(raw_query))
     assert not query.result_queued
+
+
+@pytest.mark.asyncio
+@pytest.mark.timeout(60)
+async def test_cancel(
+    *,
+    app: FastAPI,
+    kafka_broker: KafkaBroker,
+    kafka_status_consumer: AIOKafkaConsumer,
+    redis: RedisContainer,
+) -> None:
+    async with LifespanManager(app):
+        await start_query(kafka_broker, "simple")
+        status = await wait_for_status(kafka_status_consumer, "simple-started")
+        assert status.query_info
+        start_time = status.query_info.start_time
+
+        sent_time = datetime.now(tz=UTC)
+        cancel_json = read_test_job_cancel("simple").model_dump(mode="json")
+        await kafka_broker.publish(cancel_json, config.job_cancel_topic)
+
+        status = await wait_for_status(kafka_status_consumer, "simple-aborted")
+        assert status.query_info
+        assert status.query_info.start_time == start_time
+        assert status.query_info.end_time
+        assert sent_time <= status.query_info.end_time <= datetime.now(tz=UTC)
+
+    # Ensure all query state has been deleted.
+    redis_client = redis.get_client()
+    assert set(redis_client.scan_iter("query:*")) == set()
+
+
+@pytest.mark.asyncio
+@pytest.mark.timeout(60)
+async def test_upload(
+    *,
+    app: FastAPI,
+    kafka_broker: KafkaBroker,
+    kafka_status_consumer: AIOKafkaConsumer,
+    mock_qserv: MockQserv,
+    redis: RedisContainer,
+) -> None:
+    async with LifespanManager(app):
+        factory = context_dependency.create_factory()
+
+        job = await start_query(kafka_broker, "upload")
+        table_name = job.upload_tables[0].table_name
+        status = await wait_for_status(kafka_status_consumer, "upload-started")
+        assert status.query_info
+        start_time = status.query_info.start_time
+        assert mock_qserv.get_uploaded_table() == table_name
+
+        await mock_qserv.update_status(
+            1,
+            AsyncQueryStatus(
+                query_id=1,
+                status=AsyncQueryPhase.FAILED,
+                total_chunks=10,
+                completed_chunks=8,
+                collected_bytes=200,
+                query_begin=start_time,
+                last_update=datetime.now(tz=UTC),
+            ),
+        )
+        await wait_for_dispatch(factory, 1)
+
+    # Before the backend worker runs, the table should still exist.
+    assert mock_qserv.get_uploaded_table() == table_name
+
+    # Run the backend worker.
+    assert await run_arq_jobs() == 1
+    await wait_for_status(kafka_status_consumer, "upload-failed")
+
+    # Now that results have been processed, the table should be deleted.
+    assert mock_qserv.get_uploaded_table() is None
+
+    # Ensure all query state has been deleted.
+    redis_client = redis.get_client()
+    assert set(redis_client.scan_iter("query:*")) == set()
 
 
 @pytest.mark.asyncio
